# Conflicts:
#	src/Platform/SDLPlatform/SDLPlatform.cpp

+++ src/Platform/SDLPlatform/SDLInput.cpp
@@ -87,7 +87,7 @@ namespace Menge
 				WChar text_code[8];
 				size_t text_code_size;
 				UNICODE_SERVICE( m_serviceProvider )
-					->utf8ToUnicode( _event.text.text, -1, text_code, 8, &text_code_size );
+					->utf8ToUnicode( _event.text.text, (size_t)-1, text_code, 8, &text_code_size );
 
 				INPUT_SERVICE( m_serviceProvider )
 					->pushTextEvent( point.x, point.y, text_code[0] );
